+++ mayan/apps/document_indexing/api_views/index_instance_api_views.py
@@ -7,7 +7,7 @@ from mayan.apps.documents.serializers.document_serializers import DocumentSerial
 from mayan.apps.rest_api import generics
 from mayan.apps.rest_api.api_view_mixins import ExternalObjectAPIViewMixin
 
-from ..models import IndexInstance
+from ..models.index_instance_models import IndexInstance
 from ..permissions import permission_index_instance_view
 from ..serializers import (
     IndexInstanceNodeSerializer, IndexInstanceSerializer