+++ mayan/apps/document_indexing/models/__init__.py
@@ -1,0 +1,2 @@
+from .index_instance_models import *  # NOQA
+from .index_template_models import *  # NOQA