+++ mayan/apps/document_indexing/tests/test_index_template_views.py
@@ -4,7 +4,7 @@ from mayan.apps.documents.tests.base import GenericDocumentViewTestCase
 from ..events import (
     event_index_template_created, event_index_template_edited
 )
-from ..models import IndexInstanceNode, IndexTemplate
+from ..models.index_instance_models import IndexInstanceNode, IndexTemplate
 from ..permissions import (
     permission_index_template_create, permission_index_template_delete,
     permission_index_template_edit, permission_index_template_rebuild