+++ mayan/apps/tags/tests/test_indexing.py
@@ -1,5 +1,5 @@
 from mayan.apps.documents.tests.base import GenericDocumentTestCase
-from mayan.apps.document_indexing.models import IndexInstanceNode
+from mayan.apps.document_indexing.models.index_instance_models import IndexInstanceNode
 from mayan.apps.document_indexing.tests.mixins import IndexTemplateTestMixin
 
 from .literals import (